exchange JPY to ETH

+++ backend/db/migrate/20240306203432_create_coupons.rb
@@ -4,7 +4,7 @@ class CreateCoupons < ActiveRecord::Migration[7.1]
       t.string :title, null: false
       t.text :detail, null: false
       t.text :thumbnail, null: false
-      t.float :price, null: false
+      t.integer :price, null: false
       t.references :organization, null: false, foreign_key: true
 
       t.timestamps
