+++ backend/db/schema.rb
@@ -15,7 +15,7 @@ ActiveRecord::Schema[7.1].define(version: 2024_03_09_181633) do
     t.string "title", null: false
     t.text "detail", null: false
     t.text "thumbnail", null: false
-    t.float "price", null: false
+    t.integer "price", null: false
     t.integer "organization_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
